move item-check config
