+++ renderer/src/web/item-check/WidgetItemCheck.vue
@@ -14,7 +14,8 @@ import { computed, inject, ref } from 'vue'
 import { MainProcess } from '@/web/background/IPC'
 import { ItemCategory, ItemRarity, parseClipboard, ParsedItem } from '@/parser'
 import { registerActions } from './hotkeyable-actions'
-import type { ItemCheckWidget, WidgetManager } from '../overlay/interfaces'
+import type { WidgetManager } from '../overlay/interfaces'
+import type { ItemCheckWidget } from './widget.js'
 
 import Widget from '../overlay/Widget.vue'
 import MapCheck from '../map-check/MapCheck.vue'
