+++ renderer/src/web/map-check/SettingsMatcherEntry.vue
@@ -22,8 +22,7 @@
 <script lang="ts">
 import { defineComponent, computed, PropType } from 'vue'
 import { useI18n } from 'vue-i18n'
-import type { ItemCheckWidget } from '@/web/overlay/interfaces'
-import { StatMatcher, StatTag, decisionHasColor, decisionCreate } from './common.js'
+import { StatMatcher, StatTag, decisionHasColor, decisionCreate, MapCheckConfig } from './common.js'
 
 export default defineComponent({
   emits: [],
@@ -33,7 +32,7 @@ export default defineComponent({
       required: true
     },
     selectedStats: {
-      type: Array as PropType<ItemCheckWidget['maps']['selectedStats']>,
+      type: Array as PropType<MapCheckConfig['selectedStats']>,
       required: true
     },
     profile: {
