+++ renderer/src/web/settings/SettingsWindow.vue
@@ -61,7 +61,7 @@ import SettingsChat from './chat.vue'
 import SettingsGeneral from './general.vue'
 import SettingsAbout from './about.vue'
 import SettingsPricecheck from './price-check.vue'
-import SettingsItemcheck from './item-check.vue'
+import SettingsItemcheck from '../item-check/settings-item-check.vue'
 import SettingsDebug from './debug.vue'
 import SettingsMaps from '../map-check/settings-maps.vue'
 import SettingsStashSearch from '../stash-search/stash-search-editor.vue'
